API de agenda completa - Em teste

+++ controllers/clientController.js
@@ -1,9 +1,26 @@
 const { Client, User, sequelize } = require('../models');
 const bcrypt = require('bcrypt');
-const { generateUserName, generatePassword } = require('../utils/userHelpers'); // Funções auxiliares
+const { generateUserName, generatePassword } = require('../utils/userHelpers'); 
+const { validateCPF } = require('../utils/verifyingDigitHelper'); 
+const yup = require('yup');
 const dotenv = require('dotenv');
 
 dotenv.config();
+
+const clientSchema = yup.object().shape({
+  Name: yup.string().required(),
+  Email: yup.string().email().required(),
+  Phone: yup.string().optional(),
+  DateOfBirth: yup.date().nullable(),
+  Gender: yup.string().oneOf(['Male', 'Female', 'Other']).nullable(),
+  CPF: yup.string().required("CPF is mandatory")
+        .test('is-valid-cpf', 'Invalid CPF', value => validateCPF(value)),
+  Street: yup.string().nullable(),
+  Complement: yup.string().nullable(),
+  District: yup.string().nullable(),
+  ID_City: yup.number().required(),
+  CEP: yup.string().nullable(),
+});
 
 
 const clientController = {
@@ -11,8 +28,23 @@ const clientController = {
     createClient: async (req, res) => {
       const transaction = await sequelize.transaction(); // Inicia uma transação
       try {
-        const { Name, Email, Phone } = req.body;
+        await clientSchema.validate(req.body);
+ 
+        const { Name, Email, Phone, DateOfBirth, Gender, Street, Complement,
+                District, ID_City, CEP,CPF } = req.body;  
         const { ID_Company } = req.user;
+
+      // Verifica se o ID_City fornecido existe na tabela City
+      const cityExists = await City.findByPk(ID_City);
+      if (!cityExists) {
+        throw new Error('Invalid ID_City. City does not exist.');
+      }
+
+         // Verificar se o CPF já existe
+      const cpfExists = await Client.findOne({ where: { CPF }, transaction });
+      if (cpfExists) {
+        throw new Error('CPF already registered.');
+      }
 
         // Obtém o comprimento da senha do .env, ou usa 6 como padrão se não estiver definido
         const passwordLength = parseInt(process.env.AUTO_GENERATED_PASSWORD_LENGTH) || 6;
@@ -39,7 +71,7 @@ const clientController = {
           UserName: userName,
           UserEmail: Email,
           UserPassword: hashedPassword,
-          UserType: 'Other',
+          UserType: 'Client',
           ID_Company
         }, { transaction });
   
@@ -48,6 +80,14 @@ const clientController = {
           Name,
           Email,
           Phone,
+          DateOfBirth, 
+          Gender, 
+          Street, 
+          Complement,
+          District, 
+          ID_City, 
+          CEP,
+          CPF,
           ID_Company,
           UserID: newUser.UserID
         }, { transaction });
@@ -106,10 +146,33 @@ const clientController = {
   updateClient: async (req, res) => {
     const transaction = await sequelize.transaction(); // Inicia uma transação
     try {
+      await clientSchema.validate(req.body); // Validação usando o schema Yup
+
       const { id } = req.params;
-      const { Name, Email, Phone } = req.body;
+      const { Name, Email, Phone, DateOfBirth, Gender, Street, Complement,
+              District, ID_City, CEP, CPF } = req.body;
       const { ID_Company } = req.user;
-  
+
+      // Verifica se o ID_City fornecido existe na tabela City
+      const cityExists = await City.findByPk(ID_City);
+      if (!cityExists) {
+        throw new Error('Invalid ID_City. City does not exist.');
+      }
+
+      // Verificar se o CPF já existe para outro cliente
+      const cpfExists = await Client.findOne({
+        where: {
+          CPF,
+          ClientID: { [Sequelize.Op.ne]: id } // Verifica se o CPF pertence a outro cliente
+        },
+        transaction
+      });
+
+      if (cpfExists) {
+        throw new Error('CPF already registered for another client.');
+      }  
+
+
       // Encontra o cliente a ser atualizado para obter seu UserID
       const client = await Client.findOne({
         where: { ClientID: id, ID_Company },
@@ -123,7 +186,8 @@ const clientController = {
   
       // Atualiza o cliente
       await Client.update(
-        { Name, Email, Phone },
+        { Name, Email, Phone, DateOfBirth, Gender, Street, Complement,
+          District, ID_City, CEP, CPF },
         { where: { ClientID: id, ID_Company } },
         { transaction }
       );
@@ -151,29 +215,50 @@ const clientController = {
       const { id } = req.params;
       const { ID_Company } = req.user;
   
-      // Primeiro, encontra o cliente para obter o UserID associado
+      // Verifica se o cliente existe e pertence à companhia do usuário
       const client = await Client.findOne({
         where: { ClientID: id, ID_Company },
-        transaction
+          include: [
+            { 
+              model: User, 
+              as: 'clientInfo',
+              required: false 
+            },
+            {
+              model: TimeTable,
+              as: 'TimeTables', 
+              through: {
+                attributes: []
+              }
+            }
+          ],
+          transaction
       });
-  
+
       if (!client) {
         await transaction.rollback();
         return res.status(404).json({ error: 'Client not found.' });
       }
-  
-      // Deleta o cliente
+
+      if (client.TimeTables && client.TimeTables.length > 0) {
+        await transaction.rollback();
+        return res.status(400).json({ error: 'Cannot delete client because it is associated with time tables.' });
+      }
+
+      // Se o cliente tem um user associado, exclui o user
+      if (client.clientInfo) {
+        await User.destroy({
+          where: { UserID: client.clientInfo.UserID },
+          transaction
+        });
+      }
+
+      // Após remover o usuário associado, exclui o cliente
       await Client.destroy({
-        where: { ClientID: id, ID_Company },
+        where: { ClientID: id },
         transaction
       });
-  
-      // Em seguida, deleta o usuário correspondente
-      await User.destroy({
-        where: { UserID: client.UserID, ID_Company },
-        transaction
-      });
-  
+      
       await transaction.commit(); // Confirma a transação se tudo estiver correto
       res.json({ message: 'Client and corresponding user deleted successfully.' });
     } catch (error) {
